improved graphql for status, moved current_statuses on to status manager
