+++ python/aristotle-mdr-graphql/aristotle_mdr_graphql/filterset.py
@@ -47,6 +47,20 @@ class IdentifierFilterSet(FilterSet):
     class Meta:
         fields = ['namespace']
 
+
+class StatusFilterSet(FilterSet):
+    is_current = django_filters.BooleanFilter(method='filter_is_current')
+    ra = django_filters.CharFilter(name='registrationAuthority__uuid', lookup_expr='iexact', distinct=True)
+    class Meta:
+        fields = ['is_current', 'ra']
+
+    def filter_is_current(self, qs, name, value):
+        if name=="is_current" and value:
+            return qs.current()
+        else:
+            return qs
+
+
 class ConceptFilterSet(FilterSet):
 
     identifier = django_filters.CharFilter(name='identifiers__identifier', lookup_expr='iexact', distinct=True)
