+++ python/aristotle-metadata-registry/aristotle_mdr/models.py
@@ -819,29 +819,12 @@ class _concept(baseAristotleObject):
         concept_visibility_updated.send(sender=self.__class__, concept=self)
 
     def current_statuses(self, qs=None, when=timezone.now()):
+        # TODO: Look at deprecating qs argument
+        
         if qs is None:
             qs = self.statuses.all()
-        if hasattr(when, 'date'):
-            when = when.date()
 
-        states = status_filter(qs, when)
-        states = states.order_by("registrationAuthority", "-registrationDate", "-created")
-
-        from django.db import connection
-        if connection.vendor == 'postgresql':
-            states = states.distinct('registrationAuthority')
-        else:
-            current_ids = []
-            seen_ras = []
-            for s in states:
-                ra = s.registrationAuthority
-                if ra not in seen_ras:
-                    current_ids.append(s.pk)
-                    seen_ras.append(ra)
-            # We hit again so we can return this as a queryset
-            states = states.filter(pk__in=current_ids)
-
-        return states.select_related('registrationAuthority')
+        return qs.all().current(when)
 
     def get_download_items(self):
         """
